Fix floating-point overflow in intermediate calculation

+++ src/mhd/DirectInduction/mhd_iact.h
@@ -314,7 +314,7 @@ __attribute__((always_inline)) INLINE static void runner_iact_mhd_force(
   const float psi_over_ch_i = pi->mhd_data.psi_over_ch;
   const float psi_over_ch_j = pj->mhd_data.psi_over_ch;
 
-  const float permeability_inv = 1.f / mu_0;
+  const double permeability_inv = 1.f / mu_0;
 
   /* Get the kernel for hi. */
   const float hi_inv = 1.0f / hi;
@@ -634,7 +634,7 @@ __attribute__((always_inline)) INLINE static void runner_iact_nonsym_mhd_force(
   const float psi_over_ch_i = pi->mhd_data.psi_over_ch;
   const float psi_over_ch_j = pj->mhd_data.psi_over_ch;
 
-  const float permeability_inv = 1.0f / mu_0;
+  const double permeability_inv = 1.0f / mu_0;
 
   /* Get the kernel for hi. */
   const float hi_inv = 1.0f / hi;
